% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterD3.R
\name{clean_var_levels}
\alias{clean_var_levels}
\title{Clean variable levels to be valid CSS classes
https://stackoverflow.com/a/449000/249691}
\usage{
clean_var_levels(v)
}
\arguments{
\item{v}{character vector to be cleaned}
}
\description{
Clean variable levels to be valid CSS classes
https://stackoverflow.com/a/449000/249691
}

