% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterD3.R
\name{renderScatterD3}
\alias{renderScatterD3}
\title{Widget render function for use in Shiny}
\usage{
renderScatterD3(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

