% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterD3.R
\name{scatterD3}
\alias{scatterD3}
\title{<Add Title>}
\usage{
scatterD3(x, y, lab = NULL, size = NULL, labels_size = 10,
  fixed = FALSE, var = NULL, xlab = NULL, ylab = NULL, varlab = NULL,
  width = NULL, height = NULL)
}
\description{
<Add Description>
}

