% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterD3.R
\name{scatterD3Output}
\alias{scatterD3Output}
\title{Widget output function for use in Shiny}
\usage{
scatterD3Output(outputId, width = "100\%", height = "600px")
}
\description{
Widget output function for use in Shiny
}

