% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterD3-Rd.R, R/scatterD3.R
\name{scatterD3-shiny}
\alias{renderScatterD3}
\alias{scatterD3-shiny}
\alias{scatterD3Output}
\title{Shiny bindings for scatterD3 widgets}
\usage{
scatterD3Output(outputId, width = "100\%", height = "600px")

renderScatterD3(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates a scatterD3 scatter plot.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using scatterD3 widgets within Shiny
applications and interactive Rmd documents.
}

